feat : neck위젯 파일분리, 스트레칭 정보 프로그레스바 연동,
